Remove utils as it was only needed for median_cut
